+++ openslides/motions/serializers.py
@@ -1,4 +1,4 @@
-from typing import Dict
+from typing import Dict, Optional
 
 from django.db import transaction
 from django.utils.translation import ugettext as _
@@ -6,6 +6,7 @@ from django.utils.translation import ugettext as _
 from ..poll.serializers import default_votes_validator
 from ..utils.rest_api import (
     CharField,
+    DecimalField,
     DictField,
     Field,
     IntegerField,
@@ -101,12 +102,11 @@ class WorkflowSerializer(ModelSerializer):
     Serializer for motion.models.Workflow objects.
     """
     states = StateSerializer(many=True, read_only=True)
-    # The first_state is checked in the update() method
-    first_state = PrimaryKeyRelatedField(queryset=State.objects.all(), required=False)
 
     class Meta:
         model = Workflow
         fields = ('id', 'name', 'states', 'first_state',)
+        read_only_fields = ('first_state',)
 
     @transaction.atomic
     def create(self, validated_data):
@@ -126,17 +126,6 @@ class WorkflowSerializer(ModelSerializer):
         workflow.first_state = first_state
         workflow.save()
         return workflow
-
-    @transaction.atomic
-    def update(self, workflow, validated_data):
-        """
-        Check, if the first state is in the right workflow.
-        """
-        first_state = validated_data.get('first_state')
-        if first_state is not None:
-            if workflow.pk != first_state.workflow.pk:
-                raise ValidationError({'detail': 'You cannot select a state which is not in the workflow as the first state.'})
-        return super().update(workflow, validated_data)
 
 
 class MotionCommentsJSONSerializerField(Field):
@@ -212,7 +201,7 @@ class MotionPollSerializer(ModelSerializer):
     no = SerializerMethodField()
     abstain = SerializerMethodField()
     votes = DictField(
-        child=IntegerField(min_value=-2, allow_null=True),
+        child=DecimalField(max_digits=15, decimal_places=6, min_value=-2, allow_null=True),
         write_only=True)
     has_votes = SerializerMethodField()
 
@@ -238,21 +227,21 @@ class MotionPollSerializer(ModelSerializer):
 
     def get_yes(self, obj):
         try:
-            result = self.get_votes_dict(obj)['Yes']
+            result: Optional[str] = str(self.get_votes_dict(obj)['Yes'])
         except KeyError:
             result = None
         return result
 
     def get_no(self, obj):
         try:
-            result = self.get_votes_dict(obj)['No']
+            result: Optional[str] = str(self.get_votes_dict(obj)['No'])
         except KeyError:
             result = None
         return result
 
     def get_abstain(self, obj):
         try:
-            result = self.get_votes_dict(obj)['Abstain']
+            result: Optional[str] = str(self.get_votes_dict(obj)['Abstain'])
         except KeyError:
             result = None
         return result